projects/telegram-predictions-bot/requirements.txt
python-telegram-bot==21.8
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
APScheduler==3.10.4
python-dotenv==1.0.1
aiohttp==3.9.3
